Guard against environments without extended console
Fixes #1

Extended console methods like `console.groupCollapsed` and `console.groupEnd` are not available in the react-native node runtime, unless the chrome debugger is open, since the runtime is replaced with V8. This means that using `apolloLogger` will cause uncaught network errors in the link stack when it tries to call these functions.

This change neutralizes that threat by just assigning functions that will noop if the `console` functions are not available.