feat: look through definitions to find an operation type
Previously we were just reporting the first, which for large queries was
often a fragment definition, which doesn't have an operation. In most
use cases, there will be exactly one query or mutation. In the case
where there are multiple, we'll report one of them (no worse than
before!). Also - correct spelling of "elapsed"